Fix issue with switch property not turning on
With LZW45 off, using the setColor() method turns the actual device on but leaves the device's switch property set to off.  This issue is not present when using the setLevel() method.  This fix simply copies the set level functionality from the setLevel() method to the setColor() method which fixes the problem.