Update inovelli-bulb-multi-color-lzw42.groovy
Call `refresh()` instead of `refreshColor`.

Resolves an issue where switching from RGB to CT mode, while simultaneously changing level fails to update current level of bulb in Hubitat states.

For example, if the bulb is set to 2700K and 100%, then you change to [hue:34,saturation:100,level:10], it only updates the hue and saturation in the UI, failing to update the level value. This is most prominent of an issue when switching bulbs from one scene to another using the Room Lighting app.